Moved nested classes to outer scope
